+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Add RoboCop::Cop::Vicenzo::RSpec::NestedLetRedefinition #1;
 - Add RoboCop::Cop::Vicenzo::RSpec::NestedSubjectRedefinition #2;
 - Add RuboCop::Cop::Vicenzo::Rails::EnumInclusionOfValidation #3;
+- Add RuboCop::Cop::Vicenzo::RSpec::NestedContextImproperStart #4;